requirements.txt
flask==3.0.0
twilio==8.10.0
anthropic>=0.40.0
gunicorn==21.2.0
httpx>=0.25.0
supabase>=2.0.0
python-docx>=1.1.0
